+++ exporters/prometheus/exporter_test.go
@@ -38,166 +38,167 @@ func TestPrometheusExporter(t *testing.T) {
 		recordMetrics       func(ctx context.Context, meter otelmetric.Meter)
 		options             []Option
 		expectedFile        string
-		// enableEscaping is default off, which means most of these tests enable UTF-8.
-		enableEscaping      bool
+		// // enableEscaping is default off, which means most of these tests enable UTF-8.
+		// enableEscaping      bool
+		strategy            otlptranslator.TranslationStrategyOption
 		checkMetricFamilies func(t testing.TB, dtos []*dto.MetricFamily)
 	}{
-		{
-			name:           "counter",
-			expectedFile:   "testdata/counter.txt",
-			enableEscaping: true,
-			recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
-				opt := otelmetric.WithAttributes(
-					attribute.Key("A").String("B"),
-					attribute.Key("C").String("D"),
-					attribute.Key("E").Bool(true),
-					attribute.Key("F").Int(42),
-				)
-				counter, err := meter.Float64Counter(
-					"foo",
-					otelmetric.WithDescription("a simple counter"),
-					otelmetric.WithUnit("s"),
-				)
-				require.NoError(t, err)
-				counter.Add(ctx, 5, opt)
-				counter.Add(ctx, 10.3, opt)
-				counter.Add(ctx, 9, opt)
+		// {
+		// 	name:         "counter",
+		// 	expectedFile: "testdata/counter.txt",
+		// 	strategy:     otlptranslator.UnderscoreEscapingWithSuffixes,
+		// 	recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
+		// 		opt := otelmetric.WithAttributes(
+		// 			attribute.Key("A").String("B"),
+		// 			attribute.Key("C").String("D"),
+		// 			attribute.Key("E").Bool(true),
+		// 			attribute.Key("F").Int(42),
+		// 		)
+		// 		counter, err := meter.Float64Counter(
+		// 			"foo",
+		// 			otelmetric.WithDescription("a simple counter"),
+		// 			otelmetric.WithUnit("s"),
+		// 		)
+		// 		require.NoError(t, err)
+		// 		counter.Add(ctx, 5, opt)
+		// 		counter.Add(ctx, 10.3, opt)
+		// 		counter.Add(ctx, 9, opt)
 
-				attrs2 := attribute.NewSet(
-					attribute.Key("A").String("D"),
-					attribute.Key("C").String("B"),
-					attribute.Key("E").Bool(true),
-					attribute.Key("F").Int(42),
-				)
-				counter.Add(ctx, 5, otelmetric.WithAttributeSet(attrs2))
-			},
-			options: []Option{
-				WithNamespace("my.dotted.namespace"),
-				WithTranslationStrategy(otlptranslator.UnderscoreEscapingWithSuffixes),
-			},
-		},
-		{
-			name:           "counter that already has the unit suffix",
-			expectedFile:   "testdata/counter_noutf8_with_unit_suffix.txt",
-			enableEscaping: true,
-			recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
-				opt := otelmetric.WithAttributes(
-					attribute.Key("A").String("B"),
-					attribute.Key("C").String("D"),
-					attribute.Key("E").Bool(true),
-					attribute.Key("F").Int(42),
-				)
-				counter, err := meter.Float64Counter(
-					"foo.seconds",
-					otelmetric.WithDescription("a simple counter"),
-					otelmetric.WithUnit("s"),
-				)
-				require.NoError(t, err)
-				counter.Add(ctx, 5, opt)
-				counter.Add(ctx, 10.3, opt)
-				counter.Add(ctx, 9, opt)
+		// 		attrs2 := attribute.NewSet(
+		// 			attribute.Key("A").String("D"),
+		// 			attribute.Key("C").String("B"),
+		// 			attribute.Key("E").Bool(true),
+		// 			attribute.Key("F").Int(42),
+		// 		)
+		// 		counter.Add(ctx, 5, otelmetric.WithAttributeSet(attrs2))
+		// 	},
+		// 	options: []Option{
+		// 		WithNamespace("my.dotted.namespace"),
+		// 		WithTranslationStrategy(otlptranslator.UnderscoreEscapingWithSuffixes),
+		// 	},
+		// },
+		// {
+		// 	name:         "counter that already has the unit suffix",
+		// 	expectedFile: "testdata/counter_noutf8_with_unit_suffix.txt",
+		// 	strategy:     otlptranslator.UnderscoreEscapingWithSuffixes,
+		// 	recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
+		// 		opt := otelmetric.WithAttributes(
+		// 			attribute.Key("A").String("B"),
+		// 			attribute.Key("C").String("D"),
+		// 			attribute.Key("E").Bool(true),
+		// 			attribute.Key("F").Int(42),
+		// 		)
+		// 		counter, err := meter.Float64Counter(
+		// 			"foo.seconds",
+		// 			otelmetric.WithDescription("a simple counter"),
+		// 			otelmetric.WithUnit("s"),
+		// 		)
+		// 		require.NoError(t, err)
+		// 		counter.Add(ctx, 5, opt)
+		// 		counter.Add(ctx, 10.3, opt)
+		// 		counter.Add(ctx, 9, opt)
 
-				attrs2 := attribute.NewSet(
-					attribute.Key("A").String("D"),
-					attribute.Key("C").String("B"),
-					attribute.Key("E").Bool(true),
-					attribute.Key("F").Int(42),
-				)
-				counter.Add(ctx, 5, otelmetric.WithAttributeSet(attrs2))
-			},
-		},
-		{
-			name:         "counter with custom unit not tracked by ucum standards",
-			expectedFile: "testdata/counter_with_custom_unit_suffix.txt",
-			recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
-				opt := otelmetric.WithAttributes(
-					attribute.Key("A").String("B"),
-					attribute.Key("C").String("D"),
-					attribute.Key("E").Bool(true),
-					attribute.Key("F").Int(42),
-				)
-				counter, err := meter.Float64Counter(
-					"foo.dotted",
-					otelmetric.WithDescription("a simple counter"),
-					otelmetric.WithUnit("madeup"),
-				)
-				require.NoError(t, err)
-				counter.Add(ctx, 5, opt)
-				counter.Add(ctx, 10.3, opt)
-				counter.Add(ctx, 9, opt)
+		// 		attrs2 := attribute.NewSet(
+		// 			attribute.Key("A").String("D"),
+		// 			attribute.Key("C").String("B"),
+		// 			attribute.Key("E").Bool(true),
+		// 			attribute.Key("F").Int(42),
+		// 		)
+		// 		counter.Add(ctx, 5, otelmetric.WithAttributeSet(attrs2))
+		// 	},
+		// },
+		// {
+		// 	name:         "counter with custom unit not tracked by ucum standards",
+		// 	expectedFile: "testdata/counter_with_custom_unit_suffix.txt",
+		// 	recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
+		// 		opt := otelmetric.WithAttributes(
+		// 			attribute.Key("A").String("B"),
+		// 			attribute.Key("C").String("D"),
+		// 			attribute.Key("E").Bool(true),
+		// 			attribute.Key("F").Int(42),
+		// 		)
+		// 		counter, err := meter.Float64Counter(
+		// 			"foo.dotted",
+		// 			otelmetric.WithDescription("a simple counter"),
+		// 			otelmetric.WithUnit("madeup"),
+		// 		)
+		// 		require.NoError(t, err)
+		// 		counter.Add(ctx, 5, opt)
+		// 		counter.Add(ctx, 10.3, opt)
+		// 		counter.Add(ctx, 9, opt)
 
-				attrs2 := attribute.NewSet(
-					attribute.Key("A").String("D"),
-					attribute.Key("C").String("B"),
-					attribute.Key("E").Bool(true),
-					attribute.Key("F").Int(42),
-				)
-				counter.Add(ctx, 5, otelmetric.WithAttributeSet(attrs2))
-			},
-		},
-		{
-			name:         "counter with bracketed unit",
-			expectedFile: "testdata/counter_no_unit.txt",
-			recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
-				opt := otelmetric.WithAttributes(
-					attribute.Key("A").String("B"),
-					attribute.Key("C").String("D"),
-					attribute.Key("E").Bool(true),
-					attribute.Key("F").Int(42),
-				)
-				counter, err := meter.Float64Counter(
-					"foo",
-					otelmetric.WithDescription("a simple counter"),
-					otelmetric.WithUnit("{spans}"),
-				)
-				require.NoError(t, err)
-				counter.Add(ctx, 5, opt)
-				counter.Add(ctx, 10.3, opt)
-				counter.Add(ctx, 9, opt)
+		// 		attrs2 := attribute.NewSet(
+		// 			attribute.Key("A").String("D"),
+		// 			attribute.Key("C").String("B"),
+		// 			attribute.Key("E").Bool(true),
+		// 			attribute.Key("F").Int(42),
+		// 		)
+		// 		counter.Add(ctx, 5, otelmetric.WithAttributeSet(attrs2))
+		// 	},
+		// },
+		// {
+		// 	name:         "counter with bracketed unit",
+		// 	expectedFile: "testdata/counter_no_unit.txt",
+		// 	recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
+		// 		opt := otelmetric.WithAttributes(
+		// 			attribute.Key("A").String("B"),
+		// 			attribute.Key("C").String("D"),
+		// 			attribute.Key("E").Bool(true),
+		// 			attribute.Key("F").Int(42),
+		// 		)
+		// 		counter, err := meter.Float64Counter(
+		// 			"foo",
+		// 			otelmetric.WithDescription("a simple counter"),
+		// 			otelmetric.WithUnit("{spans}"),
+		// 		)
+		// 		require.NoError(t, err)
+		// 		counter.Add(ctx, 5, opt)
+		// 		counter.Add(ctx, 10.3, opt)
+		// 		counter.Add(ctx, 9, opt)
 
-				attrs2 := attribute.NewSet(
-					attribute.Key("A").String("D"),
-					attribute.Key("C").String("B"),
-					attribute.Key("E").Bool(true),
-					attribute.Key("F").Int(42),
-				)
-				counter.Add(ctx, 5, otelmetric.WithAttributeSet(attrs2))
-			},
-		},
-		{
-			name:           "counter that already has a total suffix",
-			expectedFile:   "testdata/counter.txt",
-			enableEscaping: true,
-			recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
-				opt := otelmetric.WithAttributes(
-					attribute.Key("A").String("B"),
-					attribute.Key("C").String("D"),
-					attribute.Key("E").Bool(true),
-					attribute.Key("F").Int(42),
-				)
-				counter, err := meter.Float64Counter(
-					"foo.total",
-					otelmetric.WithDescription("a simple counter"),
-					otelmetric.WithUnit("s"),
-				)
-				require.NoError(t, err)
-				counter.Add(ctx, 5, opt)
-				counter.Add(ctx, 10.3, opt)
-				counter.Add(ctx, 9, opt)
+		// 		attrs2 := attribute.NewSet(
+		// 			attribute.Key("A").String("D"),
+		// 			attribute.Key("C").String("B"),
+		// 			attribute.Key("E").Bool(true),
+		// 			attribute.Key("F").Int(42),
+		// 		)
+		// 		counter.Add(ctx, 5, otelmetric.WithAttributeSet(attrs2))
+		// 	},
+		// },
+		// {
+		// 	name:         "counter that already has a total suffix",
+		// 	expectedFile: "testdata/counter.txt",
+		// 	strategy:     otlptranslator.UnderscoreEscapingWithSuffixes,
+		// 	recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
+		// 		opt := otelmetric.WithAttributes(
+		// 			attribute.Key("A").String("B"),
+		// 			attribute.Key("C").String("D"),
+		// 			attribute.Key("E").Bool(true),
+		// 			attribute.Key("F").Int(42),
+		// 		)
+		// 		counter, err := meter.Float64Counter(
+		// 			"foo.total",
+		// 			otelmetric.WithDescription("a simple counter"),
+		// 			otelmetric.WithUnit("s"),
+		// 		)
+		// 		require.NoError(t, err)
+		// 		counter.Add(ctx, 5, opt)
+		// 		counter.Add(ctx, 10.3, opt)
+		// 		counter.Add(ctx, 9, opt)
 
-				attrs2 := attribute.NewSet(
-					attribute.Key("A").String("D"),
-					attribute.Key("C").String("B"),
-					attribute.Key("E").Bool(true),
-					attribute.Key("F").Int(42),
-				)
-				counter.Add(ctx, 5, otelmetric.WithAttributeSet(attrs2))
-			},
-			options: []Option{
-				WithNamespace("my.dotted.namespace"),
-				WithTranslationStrategy(otlptranslator.UnderscoreEscapingWithSuffixes),
-			},
-		},
+		// 		attrs2 := attribute.NewSet(
+		// 			attribute.Key("A").String("D"),
+		// 			attribute.Key("C").String("B"),
+		// 			attribute.Key("E").Bool(true),
+		// 			attribute.Key("F").Int(42),
+		// 		)
+		// 		counter.Add(ctx, 5, otelmetric.WithAttributeSet(attrs2))
+		// 	},
+		// 	options: []Option{
+		// 		WithNamespace("my.dotted.namespace"),
+		// 		WithTranslationStrategy(otlptranslator.UnderscoreEscapingWithSuffixes),
+		// 	},
+		// },
 		{
 			name:         "counter with suffixes disabled",
 			expectedFile: "testdata/counter_disabled_suffix.txt",
@@ -310,10 +311,10 @@ func TestPrometheusExporter(t *testing.T) {
 			},
 		},
 		{
-			name:           "sanitized attributes to labels",
-			expectedFile:   "testdata/sanitized_labels.txt",
-			enableEscaping: true,
-			options:        []Option{WithoutUnits()},
+			name:         "sanitized attributes to labels",
+			expectedFile: "testdata/sanitized_labels.txt",
+			strategy:     otlptranslator.UnderscoreEscapingWithSuffixes,
+			options:      []Option{WithoutUnits()},
 			recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
 				opt := otelmetric.WithAttributes(
 					// exact match, value should be overwritten
@@ -563,9 +564,9 @@ func TestPrometheusExporter(t *testing.T) {
 		{
 			name:         "counter utf-8 notranslation",
 			expectedFile: "testdata/counter_utf8_notranslation.txt",
+			strategy:     otlptranslator.NoTranslation,
 			options: []Option{
 				WithNamespace("my.dotted.namespace"),
-				WithTranslationStrategy(otlptranslator.NoTranslation),
 			},
 			recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
 				opt := otelmetric.WithAttributes(
@@ -635,11 +636,7 @@ func TestPrometheusExporter(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			ctx := context.Background()
 			registry := prometheus.NewRegistry()
-			strategy := otlptranslator.NoUTF8EscapingWithSuffixes
-			if tc.enableEscaping {
-				strategy = otlptranslator.UnderscoreEscapingWithSuffixes
-			}
-			opts := append(tc.options, WithRegisterer(registry), WithTranslationStrategy(strategy))
+			opts := append(tc.options, WithRegisterer(registry), WithTranslationStrategy(tc.strategy))
 			exporter, err := New(opts...)
 			require.NoError(t, err)